+++ packages/plugin-react/src/config/client.ts
@@ -13,7 +13,7 @@ const loadModule = require.resolve
 const getClientWebpack = (chain: WebpackChain) => {
   const { publicPath, isDev, chunkName, getOutput, cwd, useHash, chainClientConfig } = loadConfig()
   const shouldUseSourceMap = isDev || process.env.GENERATE_SOURCEMAP
-  const truePublicPath = isDev ? publicPath : `/client${publicPath}`
+  const truePublicPath = isDev ? publicPath : `${publicPath}client/`
   getBaseConfig(chain)
   chain.devtool(isDev ? 'cheap-module-source-map' : (shouldUseSourceMap ? 'source-map' : false))
   chain.entry(chunkName)