+++ packages/plugin-vue3/src/global.d.ts
@@ -7,6 +7,7 @@ declare global {
     __USE_SSR__?: IWindow['__USE_SSR__']
     __INITIAL_DATA__?: IWindow['__INITIAL_DATA__']
     STORE_CONTEXT?: IWindow['STORE_CONTEXT']
+    __VUE_APP__: any
   }
   const __isBrowser__: Boolean
 }